djangodocs/pip_requirements.txt
-e svn+http://code.djangoproject.com/svn/django/branches/releases/1.3.X#egg=Django
MySQL-python
django-annoying==0.7.6
south==0.7.3
django_compressor==0.8
django-mptt==0.4.2